refactor(reply): move pending-final custody into owner
